rename fluent_command_builder to gs1au_fluent_command_builder and publish it to our own gem repository.
this gem was developed by matt while he was working for gs1au (but now left). Since the gem was published under matt's personal account and also not been actively maintained, we decided that it would be easier to just clone the source and maintain our own copy of it.